feat: Use optional customerId instead of email and serialNumber
The setting of `S1_AGENT_CUSTOMER_ID` is entirely optional, and setting
this as your email and serial number is very much convention rather than
a requirement.

Update the module to instead accept an optional customer ID, which
can be any arbitrary string and deprecate the email and serialNumber
attributes.